requirements.txt
@"
Django==5.2.8
Pillow==10.3.0
requests==2.32.3
python-dotenv==1.0.1
"@ | Out-File -FilePath "requirements.txt" -Encoding utf8